# packages/contracts/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ozhzo Verse API",
    "description": "The Digital Operating System for Homes — Core Backend Service",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "API v1 Base Endpoint"
    }
  ],
  "paths": {
...
    "/homes/{home_id}": {
      "get": {
        "tags": ["Homes"],
        "summary": "Get home details and settings",
        "operationId": "get_home_details",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "home_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Home detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessResponse_HomeDetailDTO"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["Homes"],
        "summary": "Update home workspace settings",
        "operationId": "update_home",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "home_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateHomeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated home detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessResponse_HomeDetailDTO"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Homes"],
        "summary": "Soft-delete home workspace (Owner only)",
        "operationId": "delete_home",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "home_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Home deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSuccessResponse_MessageResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "MessageResponse": {
        "type": "object",
        "required": ["message"],
        "properties": {
          "message": { "type": "string" }
        }
      },
...
      "UpdateHomeRequest": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "nullable": true },
          "currency": { "type": "string", "nullable": true },
          "timezone": { "type": "string", "nullable": true },
          "address": { "type": "string", "nullable": true },
          "avatar_url": { "type": "string", "nullable": true }
        }
      },
...
      "HomeDetailDTO": {
        "type": "object",
        "required": ["id", "name", "currency", "timezone", "role", "members_count", "created_at"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "currency": { "type": "string" },
          "timezone": { "type": "string" },
          "address": { "type": "string", "nullable": true },
          "avatar_url": { "type": "string", "nullable": true },
          "role": { "type": "string" },
          "members_count": { "type": "integer" },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
...
      "ApiSuccessResponse_MessageResponse": {
        "type": "object",
        "required": ["success", "data"],
        "properties": {
          "success": { "type": "boolean", "default": true },
          "data": { "$ref": "#/components/schemas/MessageResponse" }
        }
      },
...
      "ApiSuccessResponse_HomeDetailDTO": {
        "type": "object",
        "required": ["success", "data"],
        "properties": {
          "success": { "type": "boolean", "default": true },
          "data": { "$ref": "#/components/schemas/HomeDetailDTO" }
        }
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}